+++ functions/CalculateScalingFactor.py
