+++ functions/GenNEMA.py
