+++ functions/GenerateCompartmentalImages.py
