+++ input/norm_map_256.nii
